fix so sizes represent window innerHeight
Using chrome.windows.update to change the window dimensions sets the outer dimensions. This means that setting a window with a URL bar and bookmarks bar visible to 1024x768 will only have a visible area of 1024x673. You can check this by logging out window.innerHeight after applying a resize.
Unfortunately the `chrome.windows` methods return window data, but all heights are the outer height. The height of a tab on the other hand is equivalent to the `window.innerHeight`. By subtracting the height of a tab from the height of the window, we can determine the height of the window chrome and update values in the sizes object accordingly. Now we get a viewable area of 1024x768 when we choose that. :)